+++ dart/lib/src/transport/transport.dart
@@ -6,6 +6,5 @@ import '../protocol.dart';
 /// A transport is in charge of sending the event/envelope either via http
 /// or caching in the disk.
 abstract class Transport {
-  Future<SentryId> sendSentryEvent(SentryEvent event);
-  Future<SentryId> sendSentryEnvelope(SentryEnvelope envelope);
+  Future<SentryId> send(SentryEnvelope envelope);
 }