+++ flutter/test/file_system_transport_test.dart
@@ -25,8 +25,11 @@ void main() {
 
     final transport = fixture.getSut(_channel);
     final event = SentryEvent();
+    final sdkVersion =
+        SdkVersion(name: 'fixture-sdkName', version: 'fixture-sdkVersion');
 
-    final sentryId = await transport.sendSentryEvent(event);
+    final envelope = SentryEnvelope.fromEvent(event, sdkVersion);
+    final sentryId = await transport.send(envelope);
 
     expect(sentryId, sentryId);
   });
@@ -37,8 +40,12 @@ void main() {
     });
 
     final transport = fixture.getSut(_channel);
+    final event = SentryEvent();
+    final sdkVersion =
+        SdkVersion(name: 'fixture-sdkName', version: 'fixture-sdkVersion');
 
-    final sentryId = await transport.sendSentryEvent(SentryEvent());
+    final envelope = SentryEnvelope.fromEvent(event, sdkVersion);
+    final sentryId = await transport.send(envelope);
 
     expect(SentryId.empty(), sentryId);
   });
@@ -53,7 +60,10 @@ void main() {
 
     final event =
         SentryEvent(message: SentryMessage('hi I am a special char ◤'));
-    await transport.sendSentryEvent(event);
+    final sdkVersion =
+        SdkVersion(name: 'fixture-sdkName', version: 'fixture-sdkVersion');
+    final envelope = SentryEnvelope.fromEvent(event, sdkVersion);
+    await transport.send(envelope);
 
     final envelopeList = arguments as List;
     final envelopeString = envelopeList.first as String;